deplog lib
